+++ gradle.properties
@@ -2,13 +2,13 @@
 org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties
-minecraft_version=1.21
-yarn_mappings=1.21+build.9
-loader_version=0.15.11
+minecraft_version=1.21.4
+yarn_mappings=1.21.4+build.8
+loader_version=0.16.14
 
 # Mod Properties
 mod_version = 1.1.1
 archives_base_name = PlayerHealthIndicators
 
 # Dependencies
-fabric_version=0.100.7+1.21
+fabric_version=0.119.3+1.21.4
