+++ src/main/java/me/andrew/healthindicators/mixin/EntityRendererAccessor.java
@@ -1,0 +1,12 @@
+package me.andrew.healthindicators.mixin;
+
+import net.minecraft.client.render.entity.EntityRenderDispatcher;
+import net.minecraft.client.render.entity.EntityRenderer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(EntityRenderer.class)
+public interface EntityRendererAccessor {
+    @Accessor("dispatcher")
+    EntityRenderDispatcher getDispatcher();
+}
