+++ src/main/resources/healthindicators.mixins.json
@@ -1,11 +1,12 @@
 {
-  "required": true,
-  "package": "me.andrew.healthindicators.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "client": [
-    "PlayerEntityRendererMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
+    "required": true,
+    "package": "me.andrew.healthindicators.mixin",
+    "compatibilityLevel": "JAVA_21",
+    "client": [
+        "EntityRendererAccessor",
+        "LivingEntityRenderMixin"
+    ],
+    "injectors": {
+        "defaultRequire": 1
   }
 }
